Add VimScriptFunctionServiceBase
